Model.Internal.Log

Documentation

data LogEntry a

Constructors

LogEntry

Fields

since, refCount :: !Int

contents :: a

data LogState a

Constructors

LogState !(FingerTree (LogEntry a) (LogEntry a)) !Int !Int !(Maybe a)

data Log a

Constructors

Log (IORef (LogState a))

newLog :: Monoid a => Bool -> IO (Log a)

Argument tracks if we're persistent or not. If persistent then oldCursors can start at the beginning otherwise we won't collect history beyond what is needed to support active cursors.

record :: Monoid a => a -> Log a -> IO ()

cursors :: Monoid a => Log a -> IO Int

Get a snapshot of the # of cursors outstanding

Utilities

watchNew :: Monoid a => FingerTree (LogEntry a) (LogEntry a) -> Int -> Int -> Maybe a -> (LogState a, Int)

watchOld :: Monoid a => FingerTree (LogEntry a) (LogEntry a) -> Int -> Int -> Maybe a -> (LogState a, Int)

Cursors

data Cursor a

Constructors

Cursor !(Log a) !(IORef Int)

newCursor :: Monoid a => Log a -> IO (Cursor a)

Subscribe to _new_ updates, but we won't get the history.

oldCursor :: Monoid a => Log a -> IO (Cursor a)

Subscribe to the oldest updates available. If we allocated our newLog as persistent this will hold everything.

data InvalidCursor

Constructors

InvalidCursor

advance :: Monoid a => Cursor a -> IO a

deleteCursor :: Monoid a => Cursor a -> IO ()

validCursor :: Cursor a -> IO Bool

We haven't called deleteCursor on this thing yet, have we?