Model.Log

Log

data Log a

newLog :: Monoid a => Bool -> IO (Log a)

Argument tracks if we're persistent or not. If persistent then oldCursors can start at the beginning otherwise we won't collect history beyond what is needed to support active cursors.

record :: Monoid a => a -> Log a -> IO ()

Cursors

data Cursor a

newCursor :: Monoid a => Log a -> IO (Cursor a)

Subscribe to _new_ updates, but we won't get the history.

oldCursor :: Monoid a => Log a -> IO (Cursor a)

Subscribe to the oldest updates available. If we allocated our newLog as persistent this will hold everything.

advance :: Monoid a => Cursor a -> IO a

deleteCursor :: Monoid a => Cursor a -> IO ()

validCursor :: Cursor a -> IO Bool

We haven't called deleteCursor on this thing yet, have we?

Exceptions

data InvalidCursor