Model.Par.Internal.Fiber

Documentation

data Worker

Constructors

Worker

Fields

ident :: !Int

pool :: !(Deque (Fiber ()))

workers :: !(MutableArray RealWorld Worker)

idlers :: !(IORef (Counted (MVar (Fiber ()))))

seed :: !(Gen RealWorld)

karma :: !(IORef Int)

fast :: !(IORef Bool)

newtype Fiber a

Constructors

Fiber

Fields

runFiber :: Worker -> IO a

schedule :: Fiber ()

Look for something to do locally, otherwise go look for work

interview :: Int -> Fiber ()

Go door to door randomly looking for work. Requires there to be at least one door to knock on.

referral :: Worker -> Fiber ()

We have a hot tip from somebody with a job opening!

idle :: Fiber ()

Give up and wait for somebody to wake us up.

defer :: Fiber () -> Fiber ()

Spawn a background task. We first put it into our job queue, and then we wake up an idler if there are any and have them try to steal it.

addKarma :: Int -> Fiber ()

If the computation ends and we have globally accumulated negative karma then somebody, somewhere, is blocked.