Model.Task

Documentation

type Task = Worker -> IO ()

data Worker

Constructors

Worker

Fields

ident :: !Int

pool :: !(Deque Task)

workers :: !(MutableArray RealWorld Worker)

idlers :: !(IORef (Counted (MVar Task)))

seed :: Gen RealWorld

schedule :: Task

this will do work stealing scheduling

interview :: Int -> Task

go door to door randomly looking for work

referral :: Worker -> Task

We have a hot tip from somebody with a job opening

idle :: Task

Give up

spawn :: Task -> Task

spawn a background task. We first put it into our job queue, and then we wake up an idler if there are any and have them try to steal it.